refactor: move data description interfaces from config.ts to schema.ts
